requirements.txt
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-mathfilters==1.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
docker==5.0.0
docopt==0.6.2
docutils==0.16
Faker==8.8.1
mysql-connector-python==8.0.23
pandas==1.2.2
path @ file:///tmp/build/80754af9/path_1614022220526/work
pathlib2 @ file:///tmp/build/80754af9/pathlib2_1607024983162/work
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pytest==6.2.2
regex @ file:///tmp/build/80754af9/regex_1606772724491/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
sentence-transformers==1.1.0
sentencepiece==0.1.95
simplegeneric==0.8.1
sympy @ file:///tmp/build/80754af9/sympy_1608137652302/work
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
text-unidecode==1.3
textblob==0.15.3
textdistance @ file:///tmp/build/80754af9/textdistance_1612461398012/work
texthero==1.0.9
tokenizers==0.10.2
virtualenv==20.4.2